+++ src/crypto/mod.rs
@@ -26,6 +26,6 @@ pub use aux::merkle_root;
 pub use reclaim::ReclaimCircuit;
 #[allow(unused_imports)]
 pub use transfer::TransferCircuit;
-pub use zkp::{manta_verify_reclaim_zkp, manta_verify_transfer_zkp, VerificationKey};
+pub use zkp::{manta_verify_reclaim_zkp, manta_verify_transfer_zkp};
 
 pub use dh::{manta_dh_dec, manta_dh_enc};
