+++ src/param.rs
@@ -119,3 +119,18 @@ pub(crate) type CommitmentParamVar =
 pub(crate) type MantaCoinCommitmentOpenVar =
 	<CommitmentSchemeVar as CommitmentGadget<CommitmentScheme, Fq>>::RandomnessVar;
 pub(crate) type MantaCoinCommitmentOutputVar = AffineVar<EdwardsParameters, FpVar<Fq>>;
+
+
+//=======================
+// wrapper of Blobs
+//=======================
+#[derive(Debug, Clone, Default, PartialEq)]
+pub struct VerificationKey {
+	pub data: &'static [u8],
+}
+
+
+#[derive(Debug, Clone, Default, PartialEq)]
+pub struct Parameter {
+	pub data: &'static [u8],
+}
