+++ src/test/frame.rs
@@ -97,8 +97,8 @@ fn test_constants_should_work() {
 	new_test_ext().execute_with(|| {
 		assert_ok!(Assets::init_asset(Origin::signed(1), 100));
 		assert_eq!(Assets::balance(1), 100);
-		let hash_param = HashParam::deserialize(HASH_PARAM_BYTES.as_ref());
-		let commit_param = CommitmentParam::deserialize(COMMIT_PARAM_BYTES.as_ref());
+		let hash_param = HashParam::deserialize(HASH_PARAM.data);
+		let commit_param = CommitmentParam::deserialize(COMMIT_PARAM.data);
 		let hash_param_checksum_local = hash_param.get_checksum();
 		let commit_param_checksum_local = commit_param.get_checksum();
 		let hash_param_checksum = HashParamChecksum::get();
@@ -114,7 +114,7 @@ fn test_mint_should_work() {
 		assert_ok!(Assets::init_asset(Origin::signed(1), 1000));
 		assert_eq!(Assets::balance(1), 1000);
 		assert_eq!(PoolBalance::get(), 0);
-		let commit_param = CommitmentParam::deserialize(COMMIT_PARAM_BYTES.as_ref());
+		let commit_param = CommitmentParam::deserialize(COMMIT_PARAM.data);
 		let mut rng = ChaCha20Rng::from_seed([3u8; 32]);
 		let mut sk = [0u8; 32];
 		rng.fill_bytes(&mut sk);
@@ -253,7 +253,7 @@ fn cannot_init_twice() {
 }
 
 fn mint_tokens(size: usize) -> Vec<(MantaCoin, MantaCoinPubInfo, MantaCoinPrivInfo)> {
-	let commit_param = CommitmentParam::deserialize(COMMIT_PARAM_BYTES.as_ref());
+	let commit_param = CommitmentParam::deserialize(COMMIT_PARAM.data);
 
 	let mut rng = ChaCha20Rng::from_seed([88u8; 32]);
 	let mut pool = 0;
@@ -303,8 +303,8 @@ fn transfer_test_helper(iter: usize) {
 	assert_eq!(Assets::balance(1), 10_000_000);
 	assert_eq!(PoolBalance::get(), 0);
 
-	let hash_param = HashParam::deserialize(HASH_PARAM_BYTES.as_ref());
-	let commit_param = CommitmentParam::deserialize(COMMIT_PARAM_BYTES.as_ref());
+	let hash_param = HashParam::deserialize(HASH_PARAM.data);
+	let commit_param = CommitmentParam::deserialize(COMMIT_PARAM.data);
 
 	// load the ZKP keys
 	let mut file = File::open("transfer_pk.bin").unwrap();
@@ -510,8 +510,8 @@ fn reclaim_test_helper(iter: usize) {
 	assert_eq!(Assets::balance(1), 10_000_000);
 	assert_eq!(PoolBalance::get(), 0);
 
-	let hash_param = HashParam::deserialize(HASH_PARAM_BYTES.as_ref());
-	let commit_param = CommitmentParam::deserialize(COMMIT_PARAM_BYTES.as_ref());
+	let hash_param = HashParam::deserialize(HASH_PARAM.data);
+	let commit_param = CommitmentParam::deserialize(COMMIT_PARAM.data);
 
 	let size = iter << 1;
 	let senders = mint_tokens(size);
